add middleware

+++ routes/web.php
@@ -22,26 +22,39 @@ Route::get('/', [ListingController::class, "index"]);
 
 
 //Show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware("auth");
 
 
 //Store listing data
-Route::post('/listings', [ListingController::class, "store"]);
+Route::post('/listings', [ListingController::class, "store"])->middleware("auth");
 
 //Show edit form
-Route::get('/listings/{listing}/edit', [ListingController::class, "edit"]);
+Route::get('/listings/{listing}/edit', [ListingController::class, "edit"])->middleware("auth");
 
 //Update listing
-Route::put("/listings/{listing}", [ListingController::class, 'update']);
+Route::put("/listings/{listing}", [ListingController::class, 'update'])->middleware("auth");
 
 //Delete listing
-Route::delete("/listings/{listing}", [ListingController::class, "destroy"]);
+Route::delete("/listings/{listing}", [ListingController::class, "destroy"])->middleware("auth");
 
 //Show register create form
-Route::get("/register", [UserController::class, "create"]);
+Route::get("/register", [UserController::class, "create"])->middleware("guest");
 
 //Create new user
 Route::post("/users", [UserController::class, 'store']);
+
+//Log user out  
+
+Route::post("/logout", [UserController::class, "logout"])->middleware("auth");
+
+//Show Login Form
+Route::get('/login', [UserController::class, "login"])->name("login")->middleware("guest");
+
+
+//Login User
+Route::post("/users/authenticate", [UserController::class, "authenticate"]);
+
+
 
 //Single Listing
 Route::get('/listings/{listing}', [ListingController::class, "show"]);
